Initial Load Tests with results
Initial tests run comparing throughput when first populating and then
loading data into levelled and eleveledb.

The tests were run in series, populating first and then loading.  The
population tests were run again in-between to try and add a roughly
even underlying volume into the stores.

The initial tests were run on on a quad core iMac with 8GB of RAM and a
fusion drive.  Due to the limited footprint of the hardware, the number
of concurrent database instances was reduced to 12, rather than the 32
in the off0the-shelf leveldb test.