Add alternative bloom
Bloom filter that can take largere keys but is still efficient to build.  Allows bloom filter to be checked without first detemrining the slot.  Also, as represents the whole SST - it could be sent to the penciller to remove the need for a message pass.

The bloom is smaller and has a worse fpr than leveled_tinybloom.  Failing the bloom check isn't so bad - due to the slot index check being relatively fast and having a very low fpr.