Ugly test hack
This test (prior tot his commit) works fine.  However, something about running it with riak `make test` causes it to fail.  The process crashes when the file:delete(F2) is called.

As the test works in isolation on R16, and also because  aprocess crashing in the real world in this stage would not be the end of the world (this whole part was added as a way of dealing with some unlikely but possible tidy-up scenarios in eqc tests), losing the test in R16 is tolerable.

So In R16 tests (which includes riak make test), the delete will no longer be called in this test.