Mas i321 checkactive (#323)
* Update rebar.config

Switch to develop-3.0 lz4

* Build and test on OTP 22

* Don't add later OTP until we can test it

* OTP 23 support

* Update FUTURE.md

Confirm previous list of needs for production readiness have now been met

* Use tag for release

* Update rebar.config

Add support for OTP 24

* Allow for caching of compaction scores

Potentially reduce the overheads of scoring each file on every run.

The change also alters the default thresholds for compaction to favour longer runs (which will tend towards greater storage efficiency).

* Clarify how the new option will be controlled in Riak

* Don't hide option

* Set max limit of 24 hours on cached score

* Improve the quality of score

Move the average towards the current score if not scoring each run.   Score from more keys to get a better score (as overheads of scoring are now better sorted by setting score_onein rather than by reducing the sample size).

* Standardise retention decision

Use the same function to decide for both scoring and compaction - and avoid the situation where somethig is scored for cmpaction, but doesnt change (which was the case previously with tombstones that were still in the ledger).

* Promote cache when scanning

When scanning over a leveled store with a helper (e.g. segment filter and last modified date range), applying the filter will speed up the query when the block index cache is available to get_slots.

If it is not available, previously the leveled_sst did not then promote the cache after it had accessed the underlying blocks.

Now the code does this, and also when the cache has all been added, it extracts the largest last modified date so that sst files older than the passed in date can be immediately dismissed

* Switch to checking on get_kvrange

In production scale testing, placing te check_modified call on get_kvrange not get_slots made the performance difference.

It should help in get_lots as well, but unable to reliably get coverage in tests with this.  So for now, will leave off until a proper test can be constructed which demonstrates any benefits.

* is_active check within fold

Not within the fold fun of the leveled_runner.

This should avoid constantly having to re-merge and filter the penciller memory when running list_buckets and hitting inactive keys

* Review feedback

Also add types to try and help avoid further confusion