Amend defaults (#367)
* Amend defaults

3.0.9 testing has used different defaults for cache_size and penciller_cache_size.  There has been no noticeable drop in the performance of Riak using these defaults, so adopting here.

The new defaults have a lower memory requirement per vnode, which is useful as recent changes and performance test results are changing the standard recommendations for ring_size.

It is now preferred to choose larger ring_sizes by default (e.g. 256  for 6 - 12 nodes, 512 for 12 - 20 nodes, 1024 for 20 +).

By choosing larger ring sizes, the benefits of scaling up a cluster will continue to make a difference even as the node count goes beyond the recommended "correct" setting.  e.g. It might be reasonable (depending on hardware choices) to grow a ring_size=512 cluster to beyond 30 nodes, yet still be relatively efficient and performant at 8 nodes.

* Comment correction